+++ prod/porter/libsrc/porter/protocol/config.h
@@ -22,10 +22,10 @@ extern "C"
 #endif /* __cplusplus */
 
     extern int config_load_global(const char *config_path, PotrGlobalConfig *global);
-    extern int config_load_service(const char *config_path, int service_id,
+    extern int config_load_service(const char *config_path, int64_t service_id,
                                    PotrServiceDef *def);
     extern int config_list_service_ids(const char *config_path,
-                                       int **ids_out, int *count_out);
+                                       int64_t **ids_out, int *count_out);
 
 #ifdef __cplusplus
 }
